Update HeroSection.module.css
I've updated the CSS to ensure the image is fully visible while maintaining its aspect ratio. Here's what the changes do:

The width is set to the smaller of:
85% of the viewport width
85% of the viewport height converted to the correct width using the 16:9 ratio
The height is set to the smaller of:
85% of the viewport width converted to the correct height using the 16:9 ratio
85% of the viewport height
This means the image will:

Always maintain its 16:9 aspect ratio
Be fully visible without cropping
Scale appropriately based on viewport size
Be centered in the container
Have enough space around it to make the tilt effect more noticeable
The image should now be properly sized and the tilt effect should be more apparent since there's space around the image to show the 3D transformation.