Update Navigation.tsx
The issue was in the scroll direction comparison. I've fixed it by changing:

From: prevScrollPos > currentScrollPos
To: currentScrollPos < prevScrollPos
Now the navigation should:

Show when you're scrolling up (to go back up the page)
Hide when you're scrolling down (to go down the page)
Always show when you're at the top of the page (currentScrollPos < 10)
Always show when the mobile menu is open
The logic is now correctly aligned with the natural scrolling behavior users expect.